+++ src/__tests__/cli-remote-diagnostics.test.ts
@@ -11,7 +11,7 @@ import fs from 'node:fs';
 import path from 'node:path';
 import { normalizeError, AppError } from '@agent-device/kernel/errors';
 import { createDaemonHttpServer } from '../daemon/server/http-server.ts';
-import { resolveSessionRequestLogPath } from '../daemon/session-store.ts';
+import { resolveSessionRequestLogPath } from '../daemon/session-artifact-paths.ts';
 import { safeSessionName } from '../daemon/session-paths.ts';
 import type { DaemonRequest, DaemonResponse } from '../daemon/daemon-request.ts';
 import { runCliCapture, type CapturedCliRun } from './cli-capture.ts';
