+++ src/daemon/__tests__/request-diagnostics-http.test.ts
@@ -4,7 +4,7 @@ import fs from 'node:fs';
 import http from 'node:http';
 import path from 'node:path';
 import { createDaemonHttpServer } from '../server/http-server.ts';
-import { resolveSessionRequestLogPath } from '../session-store.ts';
+import { resolveSessionRequestLogPath } from '../session-artifact-paths.ts';
 import { safeSessionName } from '../session-paths.ts';
 import type { DaemonResponse } from '../daemon-request.ts';
 import {
