+++ src/daemon/request-execution-scope.ts
@@ -37,11 +37,8 @@ import {
   shouldRecordEventForRequest,
 } from './session-event-log.ts';
 import type { LeaseRegistry } from './lease-registry.ts';
-import {
-  resolveSessionRequestLog,
-  resolveSessionRunnerLogPath,
-  type SessionStore,
-} from './session-store.ts';
+import { type SessionStore } from './session-store.ts';
+import { resolveSessionRequestLog, resolveSessionRunnerLogPath } from './session-artifact-paths.ts';
 import type { DaemonRequest, DaemonResponse } from './daemon-request.ts';
 import type { SessionState } from './session-state.ts';
 import { teardownSessionResources } from './session-teardown.ts';
