+++ src/daemon/server/daemon-runtime.ts
@@ -1,6 +1,7 @@
 import crypto from 'node:crypto';
 import { asAppError, AppError } from '@agent-device/kernel/errors';
-import { resolveSessionRequestLogPath, SessionStore } from '../session-store.ts';
+import { SessionStore } from '../session-store.ts';
+import { resolveSessionRequestLogPath } from '../session-artifact-paths.ts';
 import { resolveDaemonPaths, resolveDaemonServerMode } from '../config.ts';
 import { createDaemonHttpServer } from './http-server.ts';
 import { trackDownloadableArtifact } from '../artifact-tracking.ts';
