+++ src/daemon/session-lifecycle/internal/inventory.ts
@@ -27,7 +27,8 @@ import { canonicalLocalDeviceKey } from '../../device-claim-paths.ts';
 import { deviceClaimIdentity } from '../../device-claims.ts';
 import type { DaemonRequest, DaemonResponse } from '../../daemon-request.ts';
 import type { SessionRef } from '../../session-state.ts';
-import { resolveSessionRunnerLogPath, SessionStore } from '../../session-store.ts';
+import { SessionStore } from '../../session-store.ts';
+import { resolveSessionRunnerLogPath } from '../../session-artifact-paths.ts';
 import {
   requireSessionOrExplicitSelector,
   resolveCommandDevice,
